requirements.txt
APScheduler==3.10.4
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.2
django-apscheduler==0.7.0
djangorestframework==3.15.2
drf-yasg==1.21.7
idna==3.10
inflection==0.5.1
packaging==24.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.3
